+++ src/addons/RealWorld/MapCoordinator.h
@@ -10,23 +10,23 @@ class Elevation;
 
 class MapCoordinator {
     private:
-        Vec2f zeroPos;
-        Vec2f offset;
+        Vec2d zeroPos;
+        Vec2d offset;
         float gridSize;
         float SCALE_REAL_TO_WORLD = 111000.0;
         Elevation* ele = 0;
         float startElevation = 0;
 
     public:
-        MapCoordinator(Vec2f zeroPos, float gridSize);
+        MapCoordinator(Vec2d zeroPos, float gridSize);
 
-        Vec2f realToWorld(Vec2f realPosition);
-        Vec2f worldToReal(Vec2f worldPosition);
+        Vec2d realToWorld(Vec2d realPosition);
+        Vec2d worldToReal(Vec2d worldPosition);
 
         float getGridSize();
         float getElevation(float x, float y);
-        float getElevation(Vec2f v);
-        Vec2f getRealBboxPosition(Vec2f worldPosition);
+        float getElevation(Vec2d v);
+        Vec2d getRealBboxPosition(Vec2d worldPosition);
 };
 
 OSG_END_NAMESPACE;
